requirements.txt
streamlit
pillow
pandas
pdfplumber
openpyxl
reportlab
python-barcode
pyarrow
